+++ src/main/java/ch/rabanti/picoxlsx4j/Workbook.java
@@ -263,7 +263,7 @@ public class Workbook {
     }
     
     /**
-     * Adding a new Worksheet. The new worksheet will be defined as current worksheet
+     * Adding a new Worksheet
      * @param name Name of the new worksheet
      * @throws WorksheetException Thrown if the name of the worksheet already exists
      * @throws FormatException Thrown if the worksheet name contains illegal characters or is out of range (length between 1 an 31)
@@ -274,7 +274,7 @@ public class Workbook {
         {
             if (this.worksheets.get(i).getSheetName().equals(name))
             {
-                throw new WorksheetException("WorksheetNameAlreadyExistsException","The worksheet with the name '" + name + "' already exists.");
+                throw new WorksheetException("WorksheetNameAlreadxExistsException","The worksheet with the name '" + name + "' already exists.");
             }
         }
         int number = this.worksheets.size() + 1;
@@ -285,7 +285,7 @@ public class Workbook {
     }
     
     /**
-     * Adding a new Worksheet with a sanitizing option. The new worksheet will be defined as current worksheet
+     * Adding a new Worksheet with a sanitizing option
      * @param name Name of the new worksheet
      * @param sanitizeSheetName If true, the name of the worksheet will be sanitized automatically according to the specifications of Excel
      * @throws WorksheetException Thrown if the name of the worksheet already exists and sanitizeSheetName is false
@@ -305,7 +305,7 @@ public class Workbook {
     }
     
     /**
-     * Adding a new Worksheet. The new worksheet will be defined as current worksheet
+     * Adding a new Worksheet
      * @param worksheet Prepared worksheet object
      * @throws WorksheetException Thrown if the name of the worksheet already exists
      * @throws FormatException Thrown if the worksheet name contains illegal characters or is out of range (length between 1 an 31
@@ -316,7 +316,7 @@ public class Workbook {
         {
             if (this.worksheets.get(i).getSheetName().equals(worksheet.getSheetName()))
             {
-                throw new WorksheetException("WorksheetNameAlreadyExistsException","The worksheet with the name '" + worksheet.getSheetName() + "' already exists.");
+                throw new WorksheetException("WorksheetNameAlreadxExistsException","The worksheet with the name '" + worksheet.getSheetName() + "' already exists.");
             }
         }
         int number = this.worksheets.size() + 1;